// src/data/generated-pages.ts
export interface ContentPage {
  slug:             string;
  heroTitle?:       string;
  heroDescription?: string;
  imageSrc?:        string;
  imageAlt?:        string;
  metaTitle?:       string;
  metaDescription?: string;
  canonical?:       string;
  robots?:          string;
  keywords?:        string;
  ogTitle?:         string;
  ogDescription?:   string;
  twitterCard?:     string;
  author?:          string;
  publishedTime?:   string;
  modifiedTime?:    string;
  readTime?:        string;
  silo?:            string;
  siloParent?:      string;
  siloParentTitle?: string;
  serviceBody?:     string;
}